Rename fnet to f_net
