+++ keras_nlp/models/__init__.py
@@ -28,7 +28,7 @@ from keras_nlp.models.distil_bert.distil_bert_preprocessor import (
 from keras_nlp.models.distil_bert.distil_bert_tokenizer import (
     DistilBertTokenizer,
 )
-from keras_nlp.models.fnet.fnet_backbone import FNetBackbone
+from keras_nlp.models.f_net.f_net_backbone import FNetBackbone
 from keras_nlp.models.roberta.roberta_backbone import RobertaBackbone
 from keras_nlp.models.roberta.roberta_classifier import RobertaClassifier
 from keras_nlp.models.roberta.roberta_preprocessor import RobertaPreprocessor
